Split logic to generate card dictionary
